requirements.txt
# requirements.txt
# List of dependencies for the Azure Monitor MCP Agent project
azure-identity
azure-monitor-query
flask
click
openai>=1.0.0
python-dotenv
pytest
pandas
openpyxl
